Time: 1 ms (51.89%), Space: 39.1 MB (61.32%) - LeetHub

+++ 1975-maximum-matrix-sum/1975-maximum-matrix-sum.cpp
@@ -4,10 +4,10 @@ public:
         int n=matrix.size();
         long long minSum=0;
         int count=0;
-        int smallest=INT_MAX;
+        long long smallest=LLONG_MAX;
         for(int i=0;i<n;i++){
             for(int j=0;j<n;j++){
-                int num=abs(matrix[i][j]);
+                long long num=abs(matrix[i][j]);
                 smallest=min(smallest,num);
                 if(matrix[i][j]<0) count++;
                 minSum+=num;
